services/image_recognition/pyproject.toml
[project]
name = "image-recognition"
version = "0.1.0"
description = "Image recognition microservice (gRPC)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "kotti-he-oide team" }
]
dependencies = [
  "grpcio>=1.62",
  "grpcio-tools>=1.62",
  "boto3>=1.34",
  "opencv-python-headless>=4.9",
  "numpy>=1.26"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "httpx>=0.27",
  "ruff>=0.5",
  "mypy>=1.10"
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_configs = true
